Fix cast

+++ PeerTutoringSystem.Application/DTOs/Authentication/UpdateUserBalanceDto.cs
@@ -2,6 +2,6 @@ namespace PeerTutoringSystem.Application.DTOs.Authentication
 {
     public class UpdateUserBalanceDto
     {
-        public double AccountBalance { get; set; }
+        public decimal AccountBalance { get; set; }
     }
 }
